devided architecture even more

+++ src/Model/IModelKonta.java
@@ -3,6 +3,9 @@ package Model;
 public interface IModelKonta {
     void utworzKonto(String login, String haslo, String typ, String imie, String nazwisko, String pesel, String adres);
     Konto wyszukajKonto(int numerKonta);
-    void dodanieSrodkow(Konto konto, int kwota);
-    void odjecieSrodkow(Konto konto, int kwota);
+    void dodanieSrodkow(Konto konto, float kwota);
+    void odjecieSrodkow(Konto konto, float kwota);
+//    void zmienTypKonta(Konto konto, String wartosc);
+//    void zmienLoginKonta(Konto konto, String wartosc);
+//    void zmienHasloKonta(Konto konto, String wartosc);
 }
